# openapi.yaml
openapi: 3.1.1
info:
  title: Engineering Thesis API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    JwtBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GetEnterpriseServiceCustomServiceResponse:
      type: object
      required:
        [
          takesCustomAppointments,
          serviceId,
          name,
          description,
          location,
          timezone,
          maxDistanceKm,
          cathegory,
          price,
          currencyIso,
        ]
      properties:
        takesCustomAppointments:
          type: boolean
          default: true
          readOnly: true
        serviceId:
          type: integer
          format: int64
          example: 1234
        name:
          type: string
          example: Basic haircut
        description:
          type: string
          example: Male or female basic haircut
        location:
          $ref: "#/components/schemas/Location"
        timezone:
          $ref: "#/components/schemas/IanaTimeZone"
        maxDistanceKm:
          type: number
          format: double
        cathegory:
          type: string
        price:
          type: number
        currencyIso:
          type: string
    GetEnterpriseServiceNonCustomServiceResponse:
      type: object
      required:
        [
          takesCustomAppointments,
          serviceId,
          name,
          description,
          location,
          timezone,
          cathegory,
          price,
          currencyIso,
        ]
      properties:
        takesCustomAppointments:
          type: boolean
          default: true
          readOnly: true
        serviceId:
          type: integer
          format: int64
          example: 1234
        name:
          type: string
          example: Basic haircut
        description:
          type: string
          example: Male or female basic haircut
        location:
          $ref: "#/components/schemas/Location"
        timezone:
          $ref: "#/components/schemas/IanaTimeZone"
        cathegory:
          type: string
        price:
          type: number
        currencyIso:
          type: string
...
    IanaTimeZone:
      type: string
      pattern: "^[A-Za-z_]+/[A-Za-z_]+$"
      example: "Europe/Warsaw"
    HoursAndMinutesTime:
      type: string
      pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
      example: "13:00"
    Location:
      type: object
      required: [address, longitude, latitude]
      properties:
        address:
          type: string
          example: ul. Nowowiejska 15/19, Warsaw, Poland
        longitude:
          type: number
          format: double
          example: 12.3248
        latitude:
          type: number
          format: double
          example: -67.224
    ServiceSearchResponseItem:
      type: object
      required:
        [
          serviceName,
          enterpriseName,
          address,
          startDatetimeGlobal,
          endDatetimeGlobal,
        ]
      properties:
        serviceName:
          type: string
        enterpriseName:
          type: string
        address:
          type: string
        startDatetimeGlobal:
          type: string
          format: date-time
        endDatetimeGlobal:
          type: string
          format: date-time
    GetServiceFreeNonCustomAppointmentsResponseItem:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
        endTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
    GetServiceFreeCustomAppointmentsResponseItem:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
        endTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
    CreateNonCustomAppointmentRequest:
      type: object
      required: [startDatetimeShopLocal, endDatetimeShopLocal]
      properties:
        startDatetimeShopLocal:
          $ref: "#/components/schemas/LocalDatetime"
        endDatetimeShopLocal:
          $ref: "#/components/schemas/LocalDatetime"
    CreateCustomAppointmentRequest:
      properties:
        startDatetimeShopLocal:
          $ref: "#/components/schemas/LocalDatetime"
        endDatetimeShopLocal:
          $ref: "#/components/schemas/LocalDatetime"
        location:
          $ref: "#/components/schemas/Location"
    LocalDatetime:
      type: string
      pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}(:\\d{2})?$"
      example: "2025-11-05T10:30:00"
...
    CreateEnterpriseServiceReviewRequest:
      type: object
      required: [numStarsOutOf5, text]
      properties:
        numStarsOutOf5:
          type: number
        text:
          type: string
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Unauthorized
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Bad request
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Forbidden request
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Internal server error
...
paths:
...
  /services/search:
    get:
      tags: [EnterpriseServices]
      operationId: searchServices
      parameters:
        - name: serviceName
          in: query
          schema:
            type: string
        - name: enterpriseName
          in: query
          schema:
            type: string
        - name: startDatetimeGlobal
          in: query
          schema:
            type: string
            format: date-time
        - name: endDatetimeGlobal
          in: query
          schema:
            type: string
            format: date-time
        - name: preferredLongitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: preferredLatitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: cathegory
          in: query
          required: true
          schema:
            type: string
        - name: maxDistanceKm
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceSearchResponseItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /services/{serviceId}/free-non-custom-appointment-slots/:
    get:
      tags: [EnterpriseServices]
      operationId: getFreeNonCustomAppointments
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dateInServiceTimezone
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetServiceFreeNonCustomAppointmentsResponseItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /services/{serviceId}/free-time-windows-for-custom-appointments:
    get:
      tags: [EnterpriseServices]
      operationId: getFreeTimeWindowsForCustomAppointments
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dateInServiceTimezone
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetServiceFreeCustomAppointmentsResponseItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /services/{serviceId}/custom-appointments-or-not:
    get:
      tags: [EnterpriseServices]
      operationId: getServiceCustomAppointmentsStatus
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: object
                required: [areCustomAppointmentsEnabled]
                properties:
                  areCustomAppointmentsEnabled:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /services/{serviceId}/non-custom-appointments/:
    post:
      tags: [Appointments]
      operationId: createNonCustomAppointment
      security:
        - JwtBearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNonCustomAppointmentRequest"
      responses:
        "201":
          description: *CREATED_DESCRIPTION
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /services/{serviceId}/custom-appointments/:
    post:
      tags: [Appointments]
      operationId: createCustomAppointment
      security:
        - JwtBearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomAppointmentRequest"
      responses:
        "201":
          description: *CREATED_DESCRIPTION
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /services/{serviceId}:
    get:
      tags: [EnterpriseServices]
      operationId: getEnterpriseService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GetEnterpriseServiceNonCustomServiceResponse"
                  - $ref: "#/components/schemas/GetEnterpriseServiceCustomServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /services/{serviceId}/reviews:
    post:
      tags: [EnterpriseServices]
      operationId: createEnterpriseServiceReview
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnterpriseServiceReviewRequest"
      responses:
        "201":
          description: *CREATED_DESCRIPTION
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...